requirements.txt
azure-cognitiveservices-speech
flask